+++ src/main/java/de/stklcode/jvault/connector/model/AppRole.java
@@ -41,6 +41,8 @@ public final class AppRole {
 
     private List<String> boundCidrList;
 
+    private List<String> secretIdBoundCidrs;
+
     private List<String> policies;
 
     @JsonProperty("secret_id_num_uses")
@@ -73,24 +75,57 @@ public final class AppRole {
     /**
      * Construct complete {@link AppRole} object.
      *
-     * @param name            Role name (required)
-     * @param id              Role ID (optional)
-     * @param bindSecretId    Bind secret ID (optional)
-     * @param boundCidrList   Whitelist of subnets in CIDR notation (optional)
-     * @param policies        List of policies (optional)
-     * @param secretIdNumUses Maximum number of uses per secret (optional)
-     * @param secretIdTtl     Maximum TTL in seconds for secrets (optional)
-     * @param tokenTtl        Token TTL in seconds (optional)
-     * @param tokenMaxTtl     Maximum token TTL in seconds, including renewals (optional)
-     * @param period          Duration in seconds, if set the token is a periodic token (optional)
+     * @param name               Role name (required)
+     * @param id                 Role ID (optional)
+     * @param bindSecretId       Bind secret ID (optional)
+     * @param secretIdBoundCidrs Whitelist of subnets in CIDR notation (optional)
+     * @param policies           List of policies (optional)
+     * @param secretIdNumUses    Maximum number of uses per secret (optional)
+     * @param secretIdTtl        Maximum TTL in seconds for secrets (optional)
+     * @param tokenTtl           Token TTL in seconds (optional)
+     * @param tokenMaxTtl        Maximum token TTL in seconds, including renewals (optional)
+     * @param period             Duration in seconds, if set the token is a periodic token (optional)
      */
-    public AppRole(final String name, final String id, final Boolean bindSecretId, final List<String> boundCidrList,
+    public AppRole(final String name, final String id, final Boolean bindSecretId, final List<String> secretIdBoundCidrs,
                    final List<String> policies, final Integer secretIdNumUses, final Integer secretIdTtl,
                    final Integer tokenTtl, final Integer tokenMaxTtl, final Integer period) {
         this.name = name;
         this.id = id;
         this.bindSecretId = bindSecretId;
+        this.secretIdBoundCidrs = secretIdBoundCidrs;
+        this.policies = policies;
+        this.secretIdNumUses = secretIdNumUses;
+        this.secretIdTtl = secretIdTtl;
+        this.tokenTtl = tokenTtl;
+        this.tokenMaxTtl = tokenMaxTtl;
+        this.period = period;
+    }
+
+    /**
+     * Construct complete {@link AppRole} object.
+     *
+     * This constructor is used for transition from {@code bound_cidr_list} to {@code secret_id_bound_cidrs} only.
+     *
+     * @param name               Role name (required)
+     * @param id                 Role ID (optional)
+     * @param bindSecretId       Bind secret ID (optional)
+     * @param boundCidrList      Whitelist of subnets in CIDR notation (optional)
+     * @param secretIdBoundCidrs Whitelist of subnets in CIDR notation (optional)
+     * @param policies           List of policies (optional)
+     * @param secretIdNumUses    Maximum number of uses per secret (optional)
+     * @param secretIdTtl        Maximum TTL in seconds for secrets (optional)
+     * @param tokenTtl           Token TTL in seconds (optional)
+     * @param tokenMaxTtl        Maximum token TTL in seconds, including renewals (optional)
+     * @param period             Duration in seconds, if set the token is a periodic token (optional)
+     */
+    AppRole(final String name, final String id, final Boolean bindSecretId, final List<String> boundCidrList,
+                   final List<String> secretIdBoundCidrs, final List<String> policies, final Integer secretIdNumUses,
+                   final Integer secretIdTtl, final Integer tokenTtl, final Integer tokenMaxTtl, final Integer period) {
+        this.name = name;
+        this.id = id;
+        this.bindSecretId = bindSecretId;
         this.boundCidrList = boundCidrList;
+        this.secretIdBoundCidrs = secretIdBoundCidrs;
         this.policies = policies;
         this.secretIdNumUses = secretIdNumUses;
         this.secretIdTtl = secretIdTtl;
@@ -122,14 +157,18 @@ public final class AppRole {
 
     /**
      * @return list of bound CIDR subnets
+     * @deprecated Use {@link #getSecretIdBoundCidrs()} instead, as this parameter is deprecated in Vault.
      */
+    @Deprecated
     public List<String> getBoundCidrList() {
         return boundCidrList;
     }
 
     /**
      * @param boundCidrList list of subnets in CIDR notation to bind role to
+     * @deprecated Use {@link #setSecretIdBoundCidrs(List)} instead, as this parameter is deprecated in Vault.
      */
+    @Deprecated
     @JsonSetter("bound_cidr_list")
     public void setBoundCidrList(final List<String> boundCidrList) {
         this.boundCidrList = boundCidrList;
@@ -137,13 +176,45 @@ public final class AppRole {
 
     /**
      * @return list of subnets in CIDR notation as comma-separated {@link String}
+     * @deprecated Use {@link #getSecretIdBoundCidrsString()} instead, as this parameter is deprecated in Vault.
      */
+    @Deprecated
     @JsonGetter("bound_cidr_list")
     @JsonInclude(JsonInclude.Include.NON_EMPTY)
     public String getBoundCidrListString() {
         if (boundCidrList == null || boundCidrList.isEmpty())
             return "";
         return String.join(",", boundCidrList);
+    }
+
+    /**
+     * @return list of bound CIDR subnets
+     * @since 0.8 replaces {@link #getBoundCidrList()}
+     */
+    public List<String> getSecretIdBoundCidrs() {
+        return secretIdBoundCidrs;
+    }
+
+    /**
+     * @param secretIdBoundCidrs List of subnets in CIDR notation to bind secrets of this role to.
+     * @since 0.8 replaces {@link #setBoundCidrList(List)}
+     */
+    @JsonSetter("secret_id_bound_cidrs")
+    public void setSecretIdBoundCidrs(final List<String> secretIdBoundCidrs) {
+        this.secretIdBoundCidrs = secretIdBoundCidrs;
+    }
+
+    /**
+     * @return List of subnets in CIDR notation as comma-separated {@link String}
+     * @since 0.8 replaces {@link #getBoundCidrListString()} ()}
+     */
+    @JsonGetter("secret_id_bound_cidrs")
+    @JsonInclude(JsonInclude.Include.NON_EMPTY)
+    public String getSecretIdBoundCidrsString() {
+        if (secretIdBoundCidrs == null || secretIdBoundCidrs.isEmpty()) {
+            return "";
+        }
+        return String.join(",", secretIdBoundCidrs);
     }
 
     /**