+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 refactoring of the internal SSL handling (#17)
 * [improvement] `VaultConnector` extends `java.io.Serializable` (#19)
 * [improvement] Added missing flags to `SealResponse` (#20)
+* [improvement] Added replication flags to `HealthResponse` (#21)
 * [improvement] Build environment and tests now compatible with Java 10
 * [dependencies] Updated dependencies to fix vulnerabilities (i.e. CVE-2018-7489)
 * [deprecation] `VaultConnectorFactory` is deprecated in favor of `VaultConnectorBuilder` with identical API (#18)